+++ waspc/data/Generator/templates/server/src/auth/providers/local/signup.ts
@@ -1,11 +1,15 @@
 {{={= =}=}}
 import { handleRejection } from '../../../utils.js'
 import { createUser } from '../../utils.js'
+import { validateAndGetAdditionalFields } from '../../utils.js'
 
 export default handleRejection(async (req, res) => {
   const userFields = req.body || {}
 
+  const additionalFields = await validateAndGetAdditionalFields(userFields)
+
   await createUser({
+    ...additionalFields,
     username: userFields.username,
     password: userFields.password,
   })