+++ waspc/e2e-test/test-outputs/waspComplexTest-golden/waspComplexTest/.wasp/out/web-app/src/auth/forms/internal/common/LoginSignupForm.tsx
@@ -1,9 +1,22 @@
-import { useContext, type FormEvent } from 'react'
-import { styled } from '../../../../stitches.config'
-import config from '../../../../config.js'
+import { useContext } from 'react'
+import { useForm, UseFormReturn } from 'react-hook-form'
 
 import { AuthContext } from '../../Auth'
-import { Form, FormInput, FormItemGroup, FormLabel, SubmitButton } from '../Form'
+import {
+  Form,
+  FormInput,
+  FormItemGroup,
+  FormLabel,
+  FormError,
+  FormTextarea,
+  SubmitButton,
+} from '../Form'
+import type {
+  AdditionalSignupFields,
+  AdditionalSignupField,
+  AdditionalSignupFieldRenderFn,
+  FormState,
+} from '../../types'
 import * as SocialIcons from '../social/SocialIcons'
 import { SocialButton } from '../social/SocialButton'
 
@@ -46,12 +59,18 @@ const SocialAuthButtons = styled('div', {
 })
 const googleSignInUrl = `${config.apiUrl}/auth/google/login`
 
+export type LoginSignupFormFields = {
+  [key: string]: string;
+}
+
 export const LoginSignupForm = ({
     state,
     socialButtonsDirection = 'horizontal',
+    additionalSignupFields,
 }: {
-    state: 'login' | 'signup',
-    socialButtonsDirection?: 'horizontal' | 'vertical';
+    state: 'login' | 'signup'
+    socialButtonsDirection?: 'horizontal' | 'vertical'
+    additionalSignupFields?: AdditionalSignupFields
 }) => {
   const {
     isLoading,
@@ -59,7 +78,10 @@ export const LoginSignupForm = ({
     setSuccessMessage,
     setIsLoading,
   } = useContext(AuthContext)
-  const cta = state === 'login' ? 'Log in' : 'Sign up';
+  const isLogin = state === 'login'
+  const cta = isLogin ? 'Log in' : 'Sign up';
+  const hookForm = useForm<LoginSignupFormFields>()
+  const { register, formState: { errors }, handleSubmit: hookFormHandleSubmit } = hookForm
 
   return (<>
         <SocialAuth>
@@ -70,4 +92,77 @@ export const LoginSignupForm = ({
           </SocialAuthButtons>
         </SocialAuth>
   </>)
+}
+
+function AdditionalFormFields({
+  hookForm,
+  formState: { isLoading },
+  additionalSignupFields,
+}: {
+  hookForm: UseFormReturn<LoginSignupFormFields>;
+  formState: FormState;
+  additionalSignupFields: AdditionalSignupFields;
+}) {
+  const {
+    register,
+    formState: { errors },
+  } = hookForm;
+
+  function renderField<ComponentType extends React.JSXElementConstructor<any>>(
+    field: AdditionalSignupField,
+    // Ideally we would use ComponentType here, but it doesn't work with react-hook-form
+    Component: any,
+    props?: React.ComponentProps<ComponentType>
+  ) {
+    return (
+      <FormItemGroup key={field.name}>
+        <FormLabel>{field.label}</FormLabel>
+        <Component
+          {...register(field.name, field.validations)}
+          {...props}
+          disabled={isLoading}
+        />
+        {errors[field.name] && (
+          <FormError>{errors[field.name].message}</FormError>
+        )}
+      </FormItemGroup>
+    );
+  }
+
+  if (areAdditionalFieldsRenderFn(additionalSignupFields)) {
+    return additionalSignupFields(hookForm, { isLoading })
+  }
+
+  return (
+    additionalSignupFields &&
+    additionalSignupFields.map((field) => {
+      if (isFieldRenderFn(field)) {
+        return field(hookForm, { isLoading })
+      }
+      switch (field.type) {
+        case 'input':
+          return renderField<typeof FormInput>(field, FormInput, {
+            type: 'text',
+          })
+        case 'textarea':
+          return renderField<typeof FormTextarea>(field, FormTextarea)
+        default:
+          throw new Error(
+            `Unsupported additional signup field type: ${field.type}`
+          )
+      }
+    })
+  )
+}
+
+function isFieldRenderFn(
+  additionalSignupField: AdditionalSignupField | AdditionalSignupFieldRenderFn
+): additionalSignupField is AdditionalSignupFieldRenderFn {
+  return typeof additionalSignupField === 'function'
+}
+
+function areAdditionalFieldsRenderFn(
+  additionalSignupFields: AdditionalSignupFields
+): additionalSignupFields is AdditionalSignupFieldRenderFn {
+  return typeof additionalSignupFields === 'function'
 }