+++ waspc/e2e-test/test-outputs/waspJob-golden/waspJob/.wasp/out/server/src/jobs/core/allJobs.js
@@ -1,5 +1,0 @@
-
-// This module exports all jobs and is imported by the server to ensure
-// any schedules that are not referenced are still loaded by NodeJS.
-
-export { MySpecialJob } from '../MySpecialJob.js'