+++ waspc/headless-test/examples/todoApp/src/client/Todo.tsx
@@ -1,5 +1,5 @@
 import React, { useState, FormEventHandler, ChangeEventHandler } from 'react'
-import { Link } from 'react-router-dom'
+import { Link } from '@wasp/router'
 
 import { useQuery } from '@wasp/queries'
 import { OptimisticUpdateDefinition, useAction } from '@wasp/actions'