+++ waspc/src/Wasp/AppSpec/ExtImport.hs
@@ -1,4 +1,6 @@
+{-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE DeriveGeneric #-}
 
 module Wasp.AppSpec.ExtImport
   ( ExtImport (..),
@@ -7,7 +9,9 @@ module Wasp.AppSpec.ExtImport
   )
 where
 
+import Data.Aeson (FromJSON, ToJSON)
 import Data.Data (Data)
+import GHC.Generics (Generic)
 import StrongPath (File', Path, Posix, Rel)
 import Wasp.AppSpec.ExternalCode (SourceExternalCodeDir)
 
@@ -28,7 +32,7 @@ data ExtImportName
     ExtImportModule Identifier
   | -- | Represents external imports like @import { Identifier } from "file.js"@
     ExtImportField Identifier
-  deriving (Show, Eq, Data)
+  deriving (Show, Eq, Data, Generic, FromJSON, ToJSON)
 
 importIdentifier :: ExtImport -> Identifier
 importIdentifier (ExtImport importName _) = case importName of