+++ waspc/src/Wasp/AppSpec/Valid.hs
@@ -60,7 +60,8 @@ validateAppSpec spec =
           validateDbIsPostgresIfPgBossUsed spec,
           validateApiRoutesAreUnique spec,
           validateApiNamespacePathsAreUnique spec,
-          validateCrudOperations spec
+          validateCrudOperations spec,
+          validatePrismaOptions spec
         ]
 
 validateExactlyOneAppExists :: AppSpec -> Maybe ValidationError
@@ -323,6 +324,45 @@ validateCrudOperations spec =
         maybeIdField = Entity.getIdField entity
         maybeIdBlockAttribute = Entity.getIdBlockAttribute entity
         (entityName, entity) = AS.resolveRef spec (AS.Crud.entity crud)
+
+validatePrismaOptions :: AppSpec -> [ValidationError]
+validatePrismaOptions spec =
+  concat
+    [ checkIfPostgresExtensionsAreUsedWithoutPostgresDbSystem,
+      checkIfDbExtensionsAreUsedWithoutPostgresDbSystem,
+      checkIfDbExtensionsAreUsedWithoutPostgresPreviewFlag
+    ]
+  where
+    checkIfPostgresExtensionsAreUsedWithoutPostgresDbSystem :: [ValidationError]
+    checkIfPostgresExtensionsAreUsedWithoutPostgresDbSystem = maybe [] check prismaClientPreviewFeatures
+      where
+        check :: [String] -> [ValidationError]
+        check previewFeatures =
+          if not isPostgresDbUsed && "postgresqlExtensions" `elem` previewFeatures
+            then [GenericValidationError "You enabled \"postgresqlExtensions\" in app.db.prisma.clientPreviewFeatures but your db system is not PostgreSQL."]
+            else []
+
+    checkIfDbExtensionsAreUsedWithoutPostgresDbSystem :: [ValidationError]
+    checkIfDbExtensionsAreUsedWithoutPostgresDbSystem = maybe [] check prismaDbExtensions
+      where
+        check :: [AS.Db.PrismaDbExtension] -> [ValidationError]
+        check value =
+          if not isPostgresDbUsed && not (null value)
+            then [GenericValidationError "If you are using app.db.prisma.dbExtensions you must use PostgreSQL as your db system."]
+            else []
+
+    checkIfDbExtensionsAreUsedWithoutPostgresPreviewFlag :: [ValidationError]
+    checkIfDbExtensionsAreUsedWithoutPostgresPreviewFlag = case (prismaDbExtensions, prismaClientPreviewFeatures) of
+      (Nothing, _) -> []
+      (Just _extensions, Just features) | "postgresqlExtensions" `elem` features -> []
+      (Just _extensions, _) -> [GenericValidationError extensionsNotEnabledMessage]
+      where
+        extensionsNotEnabledMessage = "You are using app.db.prisma.dbExtensions but you didn't enable \"postgresqlExtensions\" in app.db.prisma.clientPreviewFeatures."
+
+    isPostgresDbUsed = isPostgresUsed spec
+    prismaOptions = AS.Db.prisma =<< AS.App.db (snd $ getApp spec)
+    prismaClientPreviewFeatures = AS.Db.clientPreviewFeatures =<< prismaOptions
+    prismaDbExtensions = AS.Db.dbExtensions =<< prismaOptions
 
 -- | This function assumes that @AppSpec@ it operates on was validated beforehand (with @validateAppSpec@ function).
 -- TODO: It would be great if we could ensure this at type level, but we decided that was too much work for now.