+++ waspc/test/AnalyzerTest.hs
@@ -50,6 +50,9 @@ spec_Analyzer = do
                 "    userEntity: User,",
                 "    methods: { usernameAndPassword: {} },",
                 "    onAuthFailedRedirectTo: \"/\",",
+                "    signup: {",
+                "      additionalFields: import { fields } from \"@server/auth/signup.js\",",
+                "    },",
                 "  },",
                 "  dependencies: [",
                 "    (\"redux\", \"^4.0.5\")",
@@ -65,7 +68,8 @@ spec_Analyzer = do
                 "    system: PostgreSQL,",
                 "    seeds: [ import { devSeedSimple } from \"@server/dbSeeds.js\" ],",
                 "    prisma: {",
-                "      clientPreviewFeatures: [\"extendedWhereUnique\"]",
+                "      clientPreviewFeatures: [\"extendedWhereUnique\"],",
+                "      dbExtensions: [{ name: \"pg_trgm\", version: \"1.0.0\" }]",
                 "    }",
                 "  },",
                 "  emailSender: {",
@@ -134,6 +138,12 @@ spec_Analyzer = do
                         Auth.Auth
                           { Auth.userEntity = Ref "User" :: Ref Entity,
                             Auth.externalAuthEntity = Nothing,
+                            Auth.signup =
+                              Just $
+                                Auth.SignupOptions
+                                  { Auth.additionalFields =
+                                      Just $ ExtImport (ExtImportField "fields") (fromJust $ SP.parseRelFileP "auth/signup.js")
+                                  },
                             Auth.methods =
                               Auth.AuthMethods
                                 { Auth.usernameAndPassword = Just Auth.usernameAndPasswordConfig,
@@ -181,7 +191,8 @@ spec_Analyzer = do
                             Db.prisma =
                               Just
                                 Db.PrismaOptions
-                                  { clientPreviewFeatures = Just ["extendedWhereUnique"]
+                                  { clientPreviewFeatures = Just ["extendedWhereUnique"],
+                                    dbExtensions = Just [Db.PrismaDbExtension {Db.name = "pg_trgm", Db.version = Just "1.0.0", Db.map = Nothing, Db.schema = Nothing}]
                                   }
                           },
                     App.emailSender =