+++ waspc/waspls/src/Wasp/LSP/ExtImport/Path.hs
@@ -3,26 +3,32 @@
 
 module Wasp.LSP.ExtImport.Path
   ( ExtFileCachePath,
+    cachePathFile,
+    cachePathExtType,
     WaspStyleExtFilePath (WaspStyleExtFilePath),
     waspStylePathToCachePath,
     absPathToCachePath,
+    cachePathToAbsPathWithoutExt,
     cachePathToAbsPath,
     tryGetTsconfigForAbsPath,
+    ExtensionType,
+    allowedExts,
   )
 where
 
 import Control.Applicative ((<|>))
 import Control.Monad.IO.Class (MonadIO (liftIO))
 import Data.Hashable (Hashable (hashWithSalt))
-import Data.List (stripPrefix)
+import Data.List (isPrefixOf, stripPrefix)
 import GHC.Generics (Generic)
-import qualified Language.LSP.Server as LSP
 import qualified Path as P
 import qualified StrongPath as SP
 import qualified StrongPath.Path as SP
 import Wasp.AppSpec.ExternalCode (SourceExternalCodeDir)
+import Wasp.LSP.ServerMonads.HasProjectRootDir (HasProjectRootDir (getProjectRootDir))
 import Wasp.Project.Common (WaspProjectDir)
 import Wasp.Util.IO (doesFileExist)
+import qualified Wasp.Util.StrongPath as SP
 
 data ExtensionlessExtFile
 
@@ -30,8 +36,10 @@ data ExtensionlessExtFile
 --
 -- It is stored relative to @src/@ and without an extension so that cache lookups
 -- (starting with a 'WaspStyleExtFilePath') are more efficient.
-data ExtFileCachePath
-  = ExtFileCachePath !(SP.Path' (SP.Rel SourceExternalCodeDir) (SP.File ExtensionlessExtFile)) !ExtensionType
+data ExtFileCachePath = ExtFileCachePath
+  { cachePathFile :: !(SP.Path' (SP.Rel SourceExternalCodeDir) (SP.File ExtensionlessExtFile)),
+    cachePathExtType :: !ExtensionType
+  }
   deriving (Show, Eq, Generic)
 
 -- | Hashes only the path portion (ignoring the extension type). This is so that
@@ -48,37 +56,49 @@ instance Hashable ExtFileCachePath where
 newtype WaspStyleExtFilePath = WaspStyleExtFilePath String deriving (Show, Eq)
 
 waspStylePathToCachePath :: WaspStyleExtFilePath -> Maybe ExtFileCachePath
-waspStylePathToCachePath (WaspStyleExtFilePath waspStylePath) = do
-  -- Removes the @, making it a path relative to src/, and remove the extension.
-  relFile <- P.parseRelFile =<< stripPrefix "@" waspStylePath
-  let (extensionLessFile, extType) = splitExtensionType relFile
-  return $ ExtFileCachePath (SP.fromPathRelFile extensionLessFile) extType
+waspStylePathToCachePath (WaspStyleExtFilePath waspStylePath) =
+  case stripPrefix "@" waspStylePath >>= SP.parseRelFile of
+    Nothing -> Nothing
+    Just relPath ->
+      Just $ case SP.splitRelExtension relPath of
+        Nothing -> ExtFileCachePath relPath DotAnyTS
+        Just (relPathWithoutExt, ext) ->
+          if useExactExtension
+            then ExtFileCachePath relPathWithoutExt (DotExact ext)
+            else ExtFileCachePath relPathWithoutExt (widenExtension ext)
+  where
+    useExactExtension = "@client" `isPrefixOf` waspStylePath
 
-absPathToCachePath :: LSP.MonadLsp c m => SP.Path' SP.Abs SP.File' -> m (Maybe ExtFileCachePath)
+absPathToCachePath :: HasProjectRootDir m => SP.Path' SP.Abs (SP.File a) -> m (Maybe ExtFileCachePath)
 absPathToCachePath absFile = do
   -- Makes the path relative to src/ and deletes the extension.
-  maybeProjectDir <- (>>= SP.parseAbsDir) <$> LSP.getRootPath
+  maybeProjectDir <- getProjectRootDir
   case maybeProjectDir of
     Nothing -> pure Nothing
     Just (projectRootDir :: SP.Path' SP.Abs (SP.Dir WaspProjectDir)) ->
       let srcDir = projectRootDir SP.</> srcDirInProjectRootDir
-       in case P.stripProperPrefix (SP.toPathAbsDir srcDir) (SP.toPathAbsFile absFile) of
+       in case SP.stripProperPrefix srcDir absFile of
             Nothing -> pure Nothing
-            Just relFile -> do
-              let (extensionLessFile, extType) = splitExtensionType relFile
-              pure $ Just $ ExtFileCachePath (SP.fromPathRelFile extensionLessFile) extType
+            Just relFile -> case SP.splitRelExtension relFile of
+              Nothing -> pure Nothing
+              Just (fileWithoutExt, ext) -> pure $ Just $ ExtFileCachePath fileWithoutExt (DotExact ext)
 
-cachePathToAbsPath :: forall m c. LSP.MonadLsp c m => ExtFileCachePath -> m (Maybe (SP.Path' SP.Abs SP.File'))
-cachePathToAbsPath (ExtFileCachePath cachePath extType) = do
-  -- Converts to an absolute path and finds the appropriate extension.
-  maybeProjectDir <- (>>= SP.parseAbsDir) <$> LSP.getRootPath
+cachePathToAbsPathWithoutExt :: HasProjectRootDir m => ExtFileCachePath -> m (Maybe (SP.Path' SP.Abs (SP.File ExtensionlessExtFile)))
+cachePathToAbsPathWithoutExt (ExtFileCachePath cachePath _) = do
+  -- Converts to an absolute path, but does not add any extension.
+  maybeProjectDir <- getProjectRootDir
   case maybeProjectDir of
-    Nothing -> pure Nothing
+    Nothing -> return Nothing
     Just (projectRootDir :: SP.Path' SP.Abs (SP.Dir WaspProjectDir)) -> do
-      let fileWithNoExtension = projectRootDir SP.</> srcDirInProjectRootDir SP.</> cachePath
-      useFirstExtensionThatExists fileWithNoExtension $ allowedExts extType
+      return $ Just $ projectRootDir SP.</> srcDirInProjectRootDir SP.</> cachePath
+
+cachePathToAbsPath :: forall m a. (MonadIO m, HasProjectRootDir m) => ExtFileCachePath -> m (Maybe (SP.Path' SP.Abs (SP.File a)))
+cachePathToAbsPath cp@(ExtFileCachePath _ extType) =
+  cachePathToAbsPathWithoutExt cp >>= \case
+    Nothing -> return Nothing
+    Just absPathWithoutExt -> useFirstExtensionThatExists absPathWithoutExt $ allowedExts extType
   where
-    useFirstExtensionThatExists :: SP.Path' SP.Abs (SP.File ExtensionlessExtFile) -> [String] -> m (Maybe (SP.Path' SP.Abs SP.File'))
+    useFirstExtensionThatExists :: SP.Path' SP.Abs (SP.File ExtensionlessExtFile) -> [String] -> m (Maybe (SP.Path' SP.Abs (SP.File a)))
     useFirstExtensionThatExists _ [] = pure Nothing
     useFirstExtensionThatExists file (ext : exts) =
       case P.addExtension ext (SP.toPathAbsFile file) of
@@ -97,10 +117,10 @@ cachePathToAbsPath (ExtFileCachePath cachePath extType) = do
 -- config files exist.
 --
 -- IF the given path is not in either @src/@ subdirectory, returns nothing.
-tryGetTsconfigForAbsPath :: SP.Path' SP.Abs (SP.Dir WaspProjectDir) -> SP.Path' SP.Abs SP.File' -> Maybe (SP.Path' SP.Abs SP.File')
+tryGetTsconfigForAbsPath :: SP.Path' SP.Abs (SP.Dir WaspProjectDir) -> SP.Path' SP.Abs (SP.File a) -> Maybe (SP.Path' SP.Abs (SP.File a))
 tryGetTsconfigForAbsPath projectRootDir file = tsconfigPath [SP.reldir|src/client|] <|> tsconfigPath [SP.reldir|src/server|]
   where
-    tsconfigPath :: SP.Path' (SP.Rel WaspProjectDir) SP.Dir' -> Maybe (SP.Path' SP.Abs SP.File')
+    tsconfigPath :: SP.Path' (SP.Rel WaspProjectDir) SP.Dir' -> Maybe (SP.Path' SP.Abs (SP.File a))
     tsconfigPath folder =
       let absFolder = projectRootDir SP.</> folder
        in if SP.toPathAbsDir absFolder `P.isProperPrefixOf` SP.toPathAbsFile file
@@ -123,18 +143,13 @@ data ExtensionType
     DotExact !String
   deriving (Show)
 
--- | Convert an extension (like @\".js\"@) to an 'ExtensionType'.
-extensionType :: String -> ExtensionType
-extensionType ".js" = DotTJS
-extensionType ".jsx" = DotTJSX
-extensionType ext = DotExact ext
-
--- | Split a path into a path with no extension and an extension type. If the
--- file has no extension, it gives it 'DotAnyTS' extension type.
-splitExtensionType :: P.Path b P.File -> (P.Path b P.File, ExtensionType)
-splitExtensionType file = case P.splitExtension file of
-  Nothing -> (file, DotAnyTS)
-  Just (file', ext) -> (file', extensionType ext)
+-- | Here, widening an extension means @.js@ and @.jsx@ extensions get converted
+-- to 'DotTJS' and 'DotTJSX', respectively, meaning they will compare equal to
+-- @.ts@ and @.tsx@ extensions.
+widenExtension :: String -> ExtensionType
+widenExtension ".js" = DotTJS
+widenExtension ".jsx" = DotTJSX
+widenExtension ext = DotExact ext
 
 -- | The extensions that each 'ExtensionType' represents.
 allowedExts :: ExtensionType -> [String]