+++ waspc/waspls/src/Wasp/LSP/Server.hs
@@ -19,6 +19,7 @@ import qualified Language.LSP.Server as LSP
 import qualified Language.LSP.Types as LSP
 import System.Exit (ExitCode (ExitFailure), exitWith)
 import qualified System.Log.Logger
+import qualified Wasp.LSP.Commands as Commands
 import Wasp.LSP.Debouncer (newDebouncerIO)
 import Wasp.LSP.Handlers
 import Wasp.LSP.Reactor (startReactorThread)
@@ -38,9 +39,11 @@ lspServerHandlers stopReactor =
       didOpenHandler,
       didSaveHandler,
       didChangeHandler,
+      executeCommandHandler,
       completionHandler,
       signatureHelpHandler,
-      gotoDefinitionHandler
+      gotoDefinitionHandler,
+      codeActionHandler
     ]
 
 serve :: Maybe FilePath -> IO ()
@@ -76,8 +79,7 @@ serve maybeLogFile = do
         where
           runHandler :: ServerM a -> IO a
           runHandler handler =
-            LSP.runLspT env $ do
-              runRLspM stateTVar handler
+            LSP.runLspT env $ runRLspM stateTVar handler
 
   exitCode <-
     LSP.runServer
@@ -128,7 +130,8 @@ lspServerOptions =
     { LSP.textDocumentSync = Just syncOptions,
       LSP.completionTriggerCharacters = Just [':', ' '],
       LSP.signatureHelpTriggerCharacters = signatureHelpTriggerCharacters,
-      LSP.signatureHelpRetriggerCharacters = signatureHelpRetriggerCharacters
+      LSP.signatureHelpRetriggerCharacters = signatureHelpRetriggerCharacters,
+      LSP.executeCommandCommands = Just Commands.availableCommands
     }
 
 -- | Options to tell the client how to update the server about the state of text