+++ waspc/waspls/src/Wasp/LSP/Syntax.hs
@@ -3,12 +3,14 @@ module Wasp.LSP.Syntax
 
     -- | Module with utilities for working with/looking for patterns in CSTs
     lspPositionToOffset,
+    lspRangeToSpan,
     locationAtOffset,
     parentIs,
     hasLeft,
     isAtExprPlace,
     lexemeAt,
     findChild,
+    findAncestor,
     -- | Printing
     showNeighborhood,
   )
@@ -18,15 +20,29 @@ import Data.List (find, intercalate)
 import qualified Language.LSP.Types as J
 import qualified Wasp.Analyzer.Parser.CST as S
 import Wasp.Analyzer.Parser.CST.Traverse
+import Wasp.Analyzer.Parser.SourceSpan (SourceSpan (SourceSpan))
 import Wasp.LSP.Util (allP, anyP)
 
--- | @lspPositionToOffset srcString position@ returns 0-based offset from the
--- start of @srcString@ to the specified line and column.
+-- | @lspPositionToOffset srcString position@ converts @position@ into a 0-based
+-- offset from the start of @srcString@.
+--
+-- @position@ is a line/column offset into @srcString@.
 lspPositionToOffset :: String -> J.Position -> Int
 lspPositionToOffset srcString (J.Position l c) =
   let linesBefore = take (fromIntegral l) (lines srcString)
    in -- We add 1 to the length of each line to make sure to count the newline
       sum (map ((+ 1) . length) linesBefore) + fromIntegral c
+
+-- | @lspRangeToSpan srcString range@ converts the @range@ into a 'SourceSpan'.
+--
+-- The start and end positions in @range@ are line/column offsets into @srcString@,
+-- and the returned 'SourceSpan' contains a start and end 0-based offset from the
+-- start of @srcString@.
+lspRangeToSpan :: String -> J.Range -> SourceSpan
+lspRangeToSpan srcString (J.Range start end) =
+  let startOffset = lspPositionToOffset srcString start
+      endOffset = lspPositionToOffset srcString end
+   in SourceSpan startOffset endOffset
 
 -- | Move to the node containing the offset.
 --
@@ -95,6 +111,12 @@ showNeighborhood t =
 -- | Search for a child node with the matching "SyntaxKind".
 findChild :: S.SyntaxKind -> Traversal -> Maybe Traversal
 findChild skind t = find ((== skind) . kindAt) $ children t
+
+findAncestor :: S.SyntaxKind -> Traversal -> Maybe Traversal
+findAncestor skind t =
+  if kindAt t == skind
+    then Just t
+    else findAncestor skind =<< up t
 
 -- | @lexeme src traversal@
 lexemeAt :: String -> Traversal -> String